+++ Sources/Swift/SentryCrash/SentryCrashWrapper.swift
@@ -170,6 +170,13 @@ public final class SentryCrashWrapper: NSObject {
         
         deviceData["locale"] = Locale.autoupdatingCurrent.identifier
         
+        if #available(macOS 12, *) {
+            deviceData["ios_app_on_macos"] = self.processInfoWrapper.isiOSAppOnMac
+            deviceData["mac_catalyst_app"] = self.processInfoWrapper.isMacCatalystApp
+        }
+        
+        deviceData["ios_app_on_visionos"] = self.processInfoWrapper.isiOSAppOnVisionOS
+        
         // Set screen dimensions if available
         setScreenDimensions(&deviceData)
         
@@ -210,7 +217,7 @@ public final class SentryCrashWrapper: NSObject {
 
             if self.processInfoWrapper.isMacCatalystApp {
                 runtimeContext["name"] = "Mac Catalyst App"
-                runtimeContext["raw_description"] = "raw_description"
+                runtimeContext["raw_description"] = "mac-catalyst-app"
             }
         }