+++ Tests/SentryTests/SentryHubTests.swift
@@ -238,7 +238,7 @@ class SentryHubTests: XCTestCase {
         // Assert
         let runtimeContext = try XCTUnwrap (hub.scope.contextDictionary["runtime"] as? [String: String])
         XCTAssertEqual(runtimeContext["name"], "Mac Catalyst App")
-        XCTAssertEqual(runtimeContext["raw_description"], "raw_description")
+        XCTAssertEqual(runtimeContext["raw_description"], "mac-catalyst-app")
     }
 
     func testScopeNotEnriched_WhenScopeIsNil() {