+++ CHANGELOG.md
@@ -69,6 +69,7 @@ This changelog lists every breaking change. For a high-level overview and upgrad
 - Add `sentry.replay_id` attribute to logs ([#6515](https://github.com/getsentry/sentry-cocoa/pull/6515))
 - Structured Logs: Add log APIs to `Hub` and `Client` (#6737)
 - Add options to customize UserFeedback error messages (#6790)
+- Add isiOSAppOnVisionOS, isiOSAppOnMac, isMacCatalystApp to device context #6939
 
 ### Fixes